chore: Update npm lint script to run all linting tasks in parallel
The code changes involve updating the `lint` script in the `package.json` file. Previously, the script used the `npm-run-all` package to run the `lint:js` and `lint:css` tasks in parallel. However, with this change, the script has been updated to use the wildcard `lint:*` to run all linting tasks in parallel. This change simplifies the linting process and improves efficiency.

Note: The commit message has been generated based on the provided code changes and recent repository commits.